Don't error if there's no need

+++ run.sh
@@ -5,6 +5,6 @@ JAVA_ARGS="-server -XX:+UseG1GC -XX:MaxGCPauseMillis=150 -XX:+ParallelRefProcEna
 for FT in csv json
 do
   OFILE="result.${FT}"
-  rm ${OFILE}
+  [[ -f "$OFILE" ]] &&  rm ${OFILE}
   java ${JAVA_ARGS} -jar target/benchmarks.jar -rf ${FT} -rff ${OFILE} "$@"
 done
